UserConrroller & UserDTO correction
